requirements.txt
faker
pandas
strands-agents
bedrock-agentcore
bedrock-agentcore-starter-toolkit
boto3
streamlit>=1.28.0
altair>=4.0
protobuf>=3.20.0
pyarrow>=10.0.0
pydantic>=2.0.0